// src/enums/filters.ts
export enum NewsSources {
    THE_GUARDIAN = "the-guardian",
    NEW_YORK_TIMES = "nytimes",
    NEWS_API = "news-api"
};

export enum NewsPerPage {
    FIRST = '10',
    SECOND = '15',
    THIRD = '20',
    FOURTH = '25'
};

export enum NewsApiParamKeys {
    QUERY = "query",
    START_DATE = "start-date",
    END_DATE = "end-date",
    SORTED_BY = "sorted-by",
    API_KEY = "api-key",
    PAGE = "page",
    PAGE_SIZE = "page-size"
};